requirements.txt
psycopg2==2.8.6
psycopg2-binary==2.8.6
python-dotenv==0.14.0
slackbot==1.0.0
